+++ XidianNavigation/MBAlertView/MBHUDView.m
@@ -67,7 +67,7 @@
 {
     if(_bodyFont)
         return _bodyFont;
-    float size = 0;
+    CGFloat size = 0;
     [self.bodyText sizeWithFont:[UIFont boldSystemFontOfSize:26] minFontSize:6 actualFontSize:&size forWidth:self.contentRect.size.width / 1.3 lineBreakMode:NSLineBreakByTruncatingTail];
     _bodyFont = [UIFont boldSystemFontOfSize:size];
     return _bodyFont;
